Activity・Presenter修正
