+++ ChapterOne-master/app/src/main/java/jp/inc/eda/chapterone_master/presentation/todocreate/TodoCreatePresenter.java
@@ -23,18 +23,13 @@ import jp.inc.eda.chapterone_master.infra.repository.TodoRepository;
 class TodoCreatePresenter {
 
     interface Contract {
-        /**
-         * TODOの作成が完了した
-         * @param todo 作成したTODO
-         */
-        void onTodoCreated(Todo todo);
     }
 
     private AppCompatActivity activity;
     private TodoRepository repository;
     private Contract contract;
 
-    TodoCreatePresenter(AppCompatActivity activity, Contract _contract) {
+    TodoCreatePresenter(AppCompatActivity activity, TodoRepository repository, Contract _contract) {
         this.activity = activity;
 
         // DB取得
@@ -42,40 +37,16 @@ class TodoCreatePresenter {
         OrmaDatabase orma = module.provideOrma(activity);
         // TODO用のDAO生成
         TodoDao dao = new TodoDao(orma);
-        repository = new TodoRepository(dao);
+        this.repository = new TodoRepository(dao);
 
         this.contract = _contract;
-
-        FloatingActionButton fab = (FloatingActionButton) activity.findViewById(R.id.fab);
-        fab.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                String title = getViewText(R.id.title);
-                if(title.isEmpty()) {
-                    // タイトルが入力されていない
-                    Snackbar.make(v, "Please enter Todo title.", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
-                    return;
-                }
-                String description = getViewText(R.id.description);
-                if(description.isEmpty()) {
-                    // 内容が入力されていない
-                    Snackbar.make(v, "Please enter Todo description.", Snackbar.LENGTH_LONG)
-                        .setAction("Action", null).show();
-                    return;
-                }
-                // TODOを保存
-                Todo todo = new Todo();
-                todo.title = title;
-                todo.description = description;
-                todo.isChecked = false;
-                repository.create(todo);
-                contract.onTodoCreated(todo);
-            }
-        });
     }
 
-    private String getViewText(@IdRes int id) {
-        return ((TextView)activity.findViewById(id)).getText().toString();
+    void createTodo(String title, String description) {
+        Todo todo = new Todo();
+        todo.title = title;
+        todo.description = description;
+        todo.isChecked = false;
+        repository.create(todo);
     }
 }
